// client/src/types/zod/schemas/objects/task_statusesWhereInput.schema.ts
import * as z from 'zod';
import type { Prisma } from '@prisma/client';
import { IntFilterObjectSchema as IntFilterObjectSchema } from './IntFilter.schema';
import { StringFilterObjectSchema as StringFilterObjectSchema } from './StringFilter.schema';
import { StringNullableFilterObjectSchema as StringNullableFilterObjectSchema } from './StringNullableFilter.schema';
import { DateTimeNullableFilterObjectSchema as DateTimeNullableFilterObjectSchema } from './DateTimeNullableFilter.schema';
import { TasksListRelationFilterObjectSchema as TasksListRelationFilterObjectSchema } from './TasksListRelationFilter.schema'

const task_statuseswhereinputSchema = z.object({
  AND: z.union([z.lazy(() => task_statusesWhereInputObjectSchema), z.lazy(() => task_statusesWhereInputObjectSchema).array()]).optional(),
  OR: z.lazy(() => task_statusesWhereInputObjectSchema).array().optional(),
  NOT: z.union([z.lazy(() => task_statusesWhereInputObjectSchema), z.lazy(() => task_statusesWhereInputObjectSchema).array()]).optional(),
  id: z.union([z.lazy(() => IntFilterObjectSchema), z.number().int()]).optional(),
  name: z.union([z.lazy(() => StringFilterObjectSchema), z.string().max(100)]).optional(),
  description: z.union([z.lazy(() => StringNullableFilterObjectSchema), z.string()]).optional().nullable(),
  created_at: z.union([z.lazy(() => DateTimeNullableFilterObjectSchema), z.coerce.date()]).optional().nullable(),
  updated_at: z.union([z.lazy(() => DateTimeNullableFilterObjectSchema), z.coerce.date()]).optional().nullable(),
  tasks: z.lazy(() => TasksListRelationFilterObjectSchema).optional()
}).strict();
export const task_statusesWhereInputObjectSchema: z.ZodType<Prisma.task_statusesWhereInput> = task_statuseswhereinputSchema as unknown as z.ZodType<Prisma.task_statusesWhereInput>;
export const task_statusesWhereInputObjectZodSchema = task_statuseswhereinputSchema;